requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect2==0.5.0
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==4.0.3
django-allauth==0.49.0
django-ckeditor==6.2.0
django-extensions==3.1.5
django-filter==21.1
django-jalali==6.0.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
easy-thumbnails==2.8.1
flake8==4.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
jdatetime==4.0.0
Jinja2==3.0.3
kavenegar==1.1.2
lxml==4.8.0
MarkupSafe==2.1.1
mccabe==0.6.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
reportlab==3.6.8
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.3
